requirements.txt
tkinter
python-chess
pygame
Pillow
